refactor: made MillionsTable more streamlined such that any table can be composed of various columns
