+++ src/main/java/org/edu/ntnu/idatt2003/group49/millions/view/dashboard/components/OwnedStocks.java
@@ -1,20 +1,17 @@
 package org.edu.ntnu.idatt2003.group49.millions.view.dashboard.components;
 
-import javafx.beans.property.SimpleDoubleProperty;
-import javafx.beans.property.SimpleStringProperty;
-import javafx.collections.FXCollections;
-import javafx.scene.control.Button;
-import javafx.scene.control.TableCell;
-import javafx.scene.control.TableColumn;
-import javafx.scene.control.TableView;
 import javafx.scene.control.Label;
 import javafx.scene.layout.HBox;
 import javafx.scene.layout.Pane;
 import javafx.scene.layout.VBox;
 import org.edu.ntnu.idatt2003.group49.millions.controller.Navigator;
+import org.edu.ntnu.idatt2003.group49.millions.model.exchange.Share;
+import org.edu.ntnu.idatt2003.group49.millions.model.exchange.Stock;
 import org.edu.ntnu.idatt2003.group49.millions.view.MillionsView;
-import org.edu.ntnu.idatt2003.group49.millions.view.components.Table.StocksTable;
+import org.edu.ntnu.idatt2003.group49.millions.view.components.MillionsTable.OwnedSharesTable;
 
+import java.math.BigDecimal;
+import java.util.List;
 import java.util.Objects;
 
 public class OwnedStocks extends MillionsView {
@@ -38,7 +35,15 @@ public class OwnedStocks extends MillionsView {
     title.getStyleClass().add("owned-stocks-title");
     title.getChildren().add(ownedStocksLabel);
 
-    StocksTable table = new StocksTable(navigator);
+    OwnedSharesTable table = new OwnedSharesTable(navigator);
+    table.setItems(List.of(
+      new Share(
+        new Stock("NVDA", "Nvidia", new BigDecimal("100")), new BigDecimal("200"), new BigDecimal("150")
+      ),
+      new Share(
+        new Stock("APPL", "Apple", new BigDecimal("400")), new BigDecimal("123"), new BigDecimal("100")
+      )
+    ));
 
     VBox vBox = new VBox();
     vBox.getStyleClass().add("owned-stocks");
